refactor(runtime): extract state divergence module (#3051)
